+++ tests/ApiBoundaryTest.php
@@ -44,6 +44,7 @@ final class ApiBoundaryTest extends TestCase
         'MrDlef\OsQueryDigest\Http\ObservedSearch',
         'MrDlef\OsQueryDigest\Http\Ring\DigestingHandler',
         'MrDlef\OsQueryDigest\Http\SearchObserver',
+        'MrDlef\OsQueryDigest\Kind',
         'MrDlef\OsQueryDigest\LazyDigest',
         'MrDlef\OsQueryDigest\Monolog\DigestProcessor',
         'MrDlef\OsQueryDigest\Monolog\SafeDigest',